// iot-service/iot-angular/src/sdk/core/models/user-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserResponseDto {

  /**
   * User avatar URL
   */
  avatarUrl?: string | null;

  /**
   * Account creation timestamp
   */
  createdAt: string;

  /**
   * User ID who created this user
   */
  createdBy?: string | null;

  /**
   * User email address
   */
  email: string;

  /**
   * Owner ID (for tenant users)
   */
  idOwner?: string | null;

  /**
   * User unique identifier
   */
  idUser: string;

  /**
   * Whether the user account is active
   */
  isActive: boolean;

  /**
   * Last login timestamp
   */
  lastLoginAt?: string | null;

  /**
   * User full name
   */
  name: string;

  /**
   * User phone number
   */
  phone?: string | null;

  /**
   * User role
   */
  role: 'admin' | 'tenant';

  /**
   * Last update timestamp
   */
  updatedAt: string;

  /**
   * User ID who last updated this user
   */
  updatedBy?: string | null;
}